Dashboard for scenarios/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.21
alembic==1.11.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==1.10.9
python-multipart==0.0.6
requests==2.31.0
jinja2==3.1.2
python-dotenv==1.0.0
httpx==0.24.1